Adjust object structure to frontend updates

+++ Models/DTOs/ProductResponseDto.cs
@@ -6,7 +6,8 @@ public class ProductResponseDto
     public string Name { get; set; } = string.Empty;
     public string? Description { get; set; }
     public decimal Price { get; set; }
-    public int Amount { get; set; }
+    public double Amount { get; set; }
     public DateTime CreatedAt { get; set; }
     public DateTime? UpdatedAt { get; set; }
-} 
+}
+
